consolidate benches/synthetic/pco_store.rs into benches/synthetic.rs
